Split Travis run into lint and build
This means that Travis will show the status of the 2 things separately.
This also allows us to run part of the build on different versions of Python.

It may be possible to split the build step further into more steps but
I'm not familiar enough with the project yet to know.