Setup user tracking before creating an Org
Fixes: https://sentry.io/nyaruka/textit/issues/563557980

Initial fix solved only part of the problem. I was not aware that in
`super().post_save()` we create the Org and create example Flows for the
Org. The issue is that for every flow create we track the event
`nyaruka.flow_created` which causes *User Not Found* error on Intercom
service.