Retry failed IVR calls
We are introducing a throttled IVR channel enabling users to specify
maximum number of concurrent events because some providers don't
internally queue calls when all available lines get congested.

The other part of the implementation is a Flow option to retry failed
calls. Because we don't know why the call failed, we are going to retry
failed call up to 5 times. Some of failures still might fail for valid
reasons, like when a phone number is invalid or inactive.