Prevent users from choosing flow they are editing in some cases
There are two actions 'start someone else in a Flow' (A) and 'start
another flow' (B) and a ruleset 'run flow' (C). Users are allowed to
choose a flow that they are editing only in case (A). Cases (B) and
(C) are special because they will run the flow for the same contact and
we want to prevent that.

Also fixed the issue of not updating flow dependencies for case (B),
which enabled users to archive or delete flows which were referenced by
another flow. A migration script will add missing flow dependencies for
the case (B).

Added flow migration (version 11.8) that will remove actions and
rulesets that reference an invalid flow.