Prevent updates of ContactField labels when adding a field to a Flow
How to reproduce the problem:
- choose a ContactField, that is not used in the Flow already, and
change the label of a ContactField (i.e. Age -> Oldness)
- update a flow and add an `Update a Contact` action that will save any
value to the ContactField with an updated label

These actions will change the label of the ContactField to the
capitalized field 'key', in this case (Oldness -> age -> Age).

The problem is fixed by updating how we detect which fields are new and
should be created.