Ensure input is string in org.parse_number
When using input parameter of type float for Decimal we get a memory
representation of that number which is 'correct' but confusing to users.

If value is a float, the binary floating point value is losslessly
converted to its exact decimal equivalent. This conversion can often
require 53 or more digits of precision. For example,
`Decimal(float('1.1'))`
converts to
`Decimal('1.100000000000000088817841970012523233890533447265625')`.

(https://docs.python.org/3.6/library/decimal.html#decimal-objects)

Fixes: https://github.com/rapidpro/rapidpro/issues/914